Create utils folder
